+++ units/src/commonMain/kotlin/inkapplications/spondee/measure/Speed.kt
@@ -2,4 +2,4 @@ package inkapplications.spondee.measure
 
 import inkapplications.spondee.structure.*
 
-typealias Speed = Ratio<out Length, DurationDimension>
+typealias Speed = Ratio<Length, DurationDimension>